refactor: moved config error to error file
