+++ src/config.rs
@@ -3,82 +3,7 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use thiserror::Error;
-
-use crate::deps::Dependency;
-
-/// Custom error type for config loading
-#[derive(Debug, Error)]
-pub enum ConfigError {
-    /// Config file not found
-    #[error("Config file not found: {0}")]
-    NotFound(PathBuf),
-
-    /// Failed to read config file
-    #[error("Failed to read config at {}: {source}", path.display())]
-    Read {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-
-    /// Failed to parse TOML
-    #[error("Failed to parse config at {}: {source}", path.display())]
-    Parse {
-        path: PathBuf,
-        #[source]
-        source: toml::de::Error,
-    },
-    #[error("Config File is missing dependencies")]
-    MissingDeps,
-
-    /// `[deps]` exists but is not a TOML table.
-    #[error("[deps] at {path} must be a table, not a {actual}", path = path.display())]
-    InvalidDepsType { path: PathBuf, actual: &'static str },
-
-    /// A dep key inside `[deps]` is empty.
-    #[error("Dep name in {path} must not be empty", path = path.display())]
-    EmptyDepName { path: PathBuf },
-
-    /// A dep key contains characters outside `[a-zA-Z0-9_.-]`.
-    #[error(
-        "Dep name '{name}' in {path} contains invalid characters; allowed: a-z, A-Z, 0-9, '_', '.', '-'",
-        path = path.display()
-    )]
-    InvalidDepName { name: String, path: PathBuf },
-
-    /// A dep's update command is empty or whitespace-only.
-    #[error(
-        "Update command for dep '{name}' in {path} must not be empty",
-        path = path.display()
-    )]
-    EmptyUpdateCommand { name: String, path: PathBuf },
-
-    /// Config file already exists at the init target path.
-    #[error("Config file already exists at {0}")]
-    ConfigAlreadyExists(PathBuf),
-
-    /// Failed to create parent directory for config.
-    #[error("Failed to create config directory: {source}")]
-    ConfigDirCreate {
-        #[source]
-        source: std::io::Error,
-    },
-
-    /// Failed to write config file during init.
-    #[error("Failed to write config file: {source}")]
-    ConfigWrite {
-        #[source]
-        source: std::io::Error,
-    },
-
-    /// $HOME environment variable is not set.
-    #[error("HOME directory not set")]
-    MissingHomeDir,
-
-    #[error("Value for key '{key}' at {path} must be a quoted string")]
-    InvalidValue { path: PathBuf, key: String },
-}
+use crate::{deps::Dependency, error::ConfigError};
 
 #[derive(Debug)]
 pub struct Config {
